unref-group allows waiting in src/transport/spread/{protocol,connection}.lisp
refs #2317

Like joining a Spread group, the operation of leaving a Spread group can
only be considered complete once the connection that did the leaving
receives a membership message corresponding to the leave
operation. Since it is sometimes necessary to wait for the leaving of a
group, unref-group now supports waiting for the operation's completion.

* src/transport/spread/protocol.lisp (unref-group): added waitable?
  keyword parameter; extended documentation string accordingly

* src/transport/spread/connection.lisp (make-hook-promise):
  (ref-group connection string):
  (unref-group connection string):