Fixed race condition in scope interning in src/scope.lisp
The interned? slot of newly interned scopes was mutated outside of the
critical section. This was probably not a problem. However, the newly
interned scope was returned without its string representation
initialized. This could lead to problems in case of parallel calls to
`scope-strings' attempting to store the computed string representation.

* src/scope.lisp (intern-scope): mutate newly interned scope, in
  particular cache the string representation, while still holding the
  lock
* test/scope.lisp (test scope-root::intern-scope/stress): new test case;
  intern scopes in parallel and check results