Fixed performance regression in src/transport/inprocess/connectors.lisp
The problem was that the `out-connector' previously accessed the scope
trie data structure via a global variable (efficient) since the
introduction of first class transports, the trie is stored in the
transport and had to be looked up through a slot access in the connector
followed by a slot access in the transport. A reference to the scope
trie is now stored in the `out-connector' making access noticeably
faster.

* src/transport/inprocess/connectors.lisp (out-connector::scope-sinks):
  new slot; cache sink trie of the transport
  (shared-initialize :after out-connector t): new method; cache scope
  trie of the connector
  (handle out-connector event): use cached scope trie