Fixed configuration-inheritance-mixin change in src/patterns/mixins.lisp
A previous change, 2ff73a9c4a9eab20b07479d5a03b832c831b9159, changed
`make-child-participant' to bind `*configuration*' to nil while
creating the child participant to prevent an arbitrary current value
from leaking into the inherited configuration.

The problem with that is that `*make-participant-hook*' is also run in
the dynamic scope of that binding, causing participant created in such
a hook to use a nil configuration.

* src/patterns/mixins.lisp
  (make-child-participant :around configuration-inheritance-mixin t t):
  emulate previous `*configuration*' binding while running
  `*make-participant-hook*'