transport.spread: Handle external termination of receiver thread
In case the Lisp systems shuts down in an unusual way, it may
terminate all threads in an arbitrary order. To handle this, the
receiver thread marks itself as aborted in case it is terminated
before its bus is shut down. The bus can then skip waiting for any
Spread messages during its own shutdown.